requirements.txt
flask==3.0.3
requests==2.32.3
msal==1.30.0
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
psutil==6.1.0